move `turn_type_from_road_geom()` out of `make`
