+++ map_model/src/map.rs
@@ -14,14 +14,13 @@ use geom::{
 use raw_map::{RawBuilding, RawMap};
 
 use crate::{
-    make::turns::turn_type_from_road_geom, osm, AmenityType, Area, AreaID, AreaType, Building,
-    BuildingID, BuildingType, CommonEndpoint, CompressedMovementID, ControlStopSign,
-    ControlTrafficSignal, DirectedRoadID, Direction, DrivingSide, ExtraPOI, Intersection,
-    IntersectionControl, IntersectionID, IntersectionKind, Lane, LaneID, LaneType, Map, MapConfig,
-    MapEdits, Movement, MovementID, OffstreetParking, OriginalRoad, ParkingLot, ParkingLotID, Path,
-    PathConstraints, PathRequest, PathV2, Pathfinder, PathfinderCaching, Position, Road,
-    RoadFilter, RoadID, RoutingParams, TransitRoute, TransitRouteID, TransitStop, TransitStopID,
-    Turn, TurnID, TurnType, Zone,
+    osm, AmenityType, Area, AreaID, AreaType, Building, BuildingID, BuildingType, CommonEndpoint,
+    CompressedMovementID, ControlStopSign, ControlTrafficSignal, DirectedRoadID, Direction,
+    DrivingSide, ExtraPOI, Intersection, IntersectionControl, IntersectionID, IntersectionKind,
+    Lane, LaneID, LaneType, Map, MapConfig, MapEdits, Movement, MovementID, OffstreetParking,
+    OriginalRoad, ParkingLot, ParkingLotID, Path, PathConstraints, PathRequest, PathV2, Pathfinder,
+    PathfinderCaching, Position, Road, RoadFilter, RoadID, RoutingParams, TransitRoute,
+    TransitRouteID, TransitStop, TransitStopID, Turn, TurnID, TurnType, Zone,
 };
 
 impl Map {
@@ -1114,4 +1113,88 @@ impl Map {
         let t_type = turn_type_from_road_geom(r1, r1_angle, r2, r2_angle, self.get_i(i), self);
         (t_type, sign_pt, r1_angle, i)
     }
+}
+
+pub fn turn_type_from_road_geom(
+    r1: &Road,
+    r1_angle: Angle,
+    r2: &Road,
+    r2_angle: Angle,
+    i: &Intersection,
+    map: &Map,
+) -> TurnType {
+    // let r1_angle = src.last_line().angle();
+    // let r2_angle = dst.first_line().angle();
+    let expected_turn_types = expected_turn_types_for_four_way(i, map);
+
+    let mut turn_type = turn_type_from_angles(r1_angle, r2_angle);
+    if turn_type == TurnType::UTurn {
+        // Lots of false positives when classifying these just based on angles. So also
+        // require the road names to match.
+
+        if r1.get_name(None) != r2.get_name(None) {
+            // Distinguish really sharp lefts/rights based on clockwiseness
+            if r1_angle.simple_shortest_rotation_towards(r2_angle) < 0.0 {
+                turn_type = TurnType::Right;
+            } else {
+                turn_type = TurnType::Left;
+            }
+        }
+    } else if let Some(expected_type) = expected_turn_types
+        .as_ref()
+        .and_then(|e| e.get(&(r1.id, r2.id)))
+    {
+        // At some 4-way intersections, roads meet at strange angles, throwing off
+        // turn_type_from_angles. Correct it based on relative ordering.
+        if turn_type != *expected_type {
+            warn!(
+                "Turn from {:?} to {:?} looks like {:?} by angle, but is {:?} by ordering",
+                r1, r2, turn_type, expected_type
+            );
+            turn_type = *expected_type;
+        }
+    }
+    turn_type
+}
+
+pub fn turn_type_from_angles(from: Angle, to: Angle) -> TurnType {
+    let diff = from.simple_shortest_rotation_towards(to);
+    // This is a pretty arbitrary parameter, but a difference of 30 degrees seems reasonable for
+    // some observed cases.
+    if diff.abs() < 30.0 {
+        TurnType::Straight
+    } else if diff.abs() > 135.0 {
+        TurnType::UTurn
+    } else if diff < 0.0 {
+        // Clockwise rotation
+        TurnType::Right
+    } else {
+        // Counter-clockwise rotation
+        TurnType::Left
+    }
+}
+
+fn expected_turn_types_for_four_way(
+    i: &Intersection,
+    map: &Map,
+) -> Option<HashMap<(RoadID, RoadID), TurnType>> {
+    let roads = i.get_sorted_incoming_roads(map);
+    if roads.len() != 4 {
+        return None;
+    }
+
+    // Just based on relative ordering around the intersection, turns (from road, to road, should
+    // have this type)
+    let mut expected_turn_types: HashMap<(RoadID, RoadID), TurnType> = HashMap::new();
+    for (offset, turn_type) in [
+        (1, TurnType::Left),
+        (2, TurnType::Straight),
+        (3, TurnType::Right),
+    ] {
+        for from_idx in 0..roads.len() {
+            let to = *abstutil::wraparound_get(&roads, (from_idx as isize) + offset);
+            expected_turn_types.insert((roads[from_idx], to), turn_type);
+        }
+    }
+    Some(expected_turn_types)
 }
